# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: dnsaur API
  version: "1.0"
  description: >
    REST API for the dnsaur DNS/DHCP server. All paths are prefixed
    /api/v1. Authenticate with either the dnsaur_session cookie (set by
    POST /auth/login) or an Authorization: Bearer <token> API token
    (see /tokens). API tokens may have scope "read" (GET/HEAD only) or
    "write" (all methods); the session cookie always has full access.
    A read-scoped token used for a POST/PUT/PATCH/DELETE request is
    rejected with 403 {"error": "read-only token"} — see the 403
    response documented on each mutating operation below. Scope is
    enforced by method, with one exception on the read side: a
    read-scoped token gets TSIG secrets redacted (see the TSIGKey
    schema's secret_redacted field), because a GET that returns a
    signing credential in the clear is not something "read-only"
    should cover. The Authorization scheme name is matched
    case-insensitively per RFC 9110 §11.1.
    There is no separate "upstreams" resource: upstream DNS servers are
    configured via the "upstreams" key in /settings (a comma-separated
    list), alongside "upstream.strategy". Each entry is plain by default
    ("1.1.1.1:53") or, with a scheme, DNS-over-TLS
    ("tls://1.1.1.1:853#cloudflare-dns.com") or DNS-over-HTTPS
    ("https://1.1.1.1/dns-query#cloudflare-dns.com") — see the "upstreams"
    key under PUT /settings for the full grammar.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dnsaur_session
      description: Session cookie set by POST /auth/login. HttpOnly, SameSite=Strict.
    bearerAuth:
      type: http
      scheme: bearer
      description: API token minted via POST /tokens, sent as "Authorization Bearer <token>".
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
  responses:
    ReadOnlyToken:
      description: >
        read-only token used for a write operation, or a cookie-authenticated
        write whose Sec-Fetch-Site is cross-site ({"error": "cross-site
        request"}) -- CSRF defence in depth behind SameSite=Strict. Bearer
        requests are never checked for the latter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyAttempts:
      description: >
        too many attempts. POST /setup and POST /auth/login share one budget
        of 10 attempts per minute per source address, after which the source
        is locked out for a minute. Retry-After carries the wait in seconds.
        Every answer costs an attempt, including 428 and a malformed body,
        so neither endpoint is a free password oracle. Behind a reverse
        proxy, set the trusted_proxies bootstrap key or every request shares
        one source address and one budget.
      headers:
        Retry-After:
          schema: { type: integer }
          description: seconds to wait before the next attempt
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /setup:
    get:
      tags: [Setup]
      security: []
      summary: Whether first-run setup is required
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  setup_required: { type: boolean }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Setup]
      security: []
      summary: Create the admin account (first run only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, minLength: 8 }
                totp_code: { type: string, description: unused on setup }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: created }
        "400":
          description: "invalid input: username required, or password must be at least 8 characters"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: setup already completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in and receive a session cookie
      description: >
        On success, sets the dnsaur_session cookie (HttpOnly, SameSite=Strict,
        30-day TTL, sliding but never past 90 days from this login). Secure
        is set when TLS terminated here, or when the request arrived from a
        network named in the trusted_proxies bootstrap setting and its
        X-Forwarded-Proto is https. If the account has TOTP enabled,
        totp_code must be supplied or the request fails with 428, and a code
        that has already logged in once is refused as 401 bad credentials
        (RFC 6238 section 5.2). If first-run setup hasn't completed yet, the
        request fails with 409 instead of being evaluated against the
        (empty) users table. Shares a 10-attempt-per-minute-per-source
        budget with POST /setup; every answer costs an attempt, the 428
        included.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                totp_code: { type: string }
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              schema: { type: string }
              description: dnsaur_session=<token>; HttpOnly; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        "400":
          description: invalid json
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: bad credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "428":
          description: totp code required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: setup required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out and clear the session cookie
      responses:
        "204":
          description: No Content
          headers:
            Set-Cookie:
              schema: { type: string }
              description: dnsaur_session cleared (Max-Age -1)
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
  /auth/password:
    post:
      tags: [Auth]
      summary: Change the signed-in account's password
      description: >
        Verifies current_password, stores a new argon2id hash, and revokes
        every other session on the account — the caller's own survives, and
        API tokens are untouched. Shares POST /auth/login's per-source
        budget of 10 attempts a minute: it runs the same argon2id
        verification and answers whether the password was right, so a
        stolen session cookie must not be an unmetered oracle for it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        "204":
          description: No Content
        "400":
          description: >
            invalid json; "current password is wrong"; or "invalid input:
            password must be at least 8 characters". A wrong current
            password is 400 rather than 401 — the request authenticated
            fine, and 401 would send the dashboard back to the login screen
            over a typo.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/sessions:
    delete:
      tags: [Auth]
      summary: Log out every other session
      description: >
        Revokes every session on this account except the one making the
        request, for a cookie the admin believes has been copied. API
        tokens are a separate credential and are not swept; they are
        revoked by name via DELETE /tokens/{id}.
      responses:
        "204":
          description: No Content
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, format: int64 }
                  username: { type: string }
                  totp_enabled: { type: boolean }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...